chore: move more public_key code to public_key module
