+++ tests/mailbox.rs
@@ -3,11 +3,11 @@ use std::sync::Arc;
 
 use object_store::config::Config;
 use object_store::consts::*;
-use object_store::datastore::PublicKey;
 use object_store::pb::{
     AckRequest, Audience, GetRequest, mailbox_service_client::MailboxServiceClient,
 };
 use object_store::proto::{AudienceUtil, ObjectResponseUtil};
+use object_store::public_key::PublicKey;
 
 use sqlx::postgres::PgPool;
 use tonic::Request;
